@@ -6,15 +6,21 @@
   "license": "GPL-3.0",
   "private": true,
   "scripts": {
-    "dev": "cross-env NODE_ENV=development nodemon server/index.js --watch server",
-    "start": "cross-env NODE_ENV=production node server/index.js",
-    "lint": "eslint --ext .js,.vue .",
+    "dev": "nuxt",
     "build": "nuxt build",
+    "start": "nuxt start",
+    "export": "nuxt export",
     "generate": "nuxt generate",
+    "lint:js": "eslint --ext .js,.vue --ignore-path .gitignore .",
+    "lint": "yarn lint:js",
     "prettier": "prettier --write '**/*.{css,scss,js,vue}'",
-    "precommit": "yarn run lint"
+    "precommit": "yarn run lint",
+    "deploy": "yarn build && yarn export"
   },
   "dependencies": {
+    "@nuxtjs/pwa": "^3.0.2",
+    "core-js": "^3.6.5",
+    "nuxt": "^2.14.5",
     "@fortawesome/fontawesome-svg-core": "^1.2.19",
     "@fortawesome/free-brands-svg-icons": "^5.9.0",
     "@fortawesome/vue-fontawesome": "^0.1.6",
@@ -29,30 +35,21 @@
     "hover.css": "^2.3.2",
     "modernizr": "^3.7.1",
     "node-sass": "^4.12.0",
-    "nuxt": "^2.11.0",
     "nuxt-webfontloader": "^1.1.0",
     "pug": "^2.0.3",
     "pug-plain-loader": "^1.0.0",
     "sass-loader": "^v8.0.0",
-    "vue-cookie-law": "^1.10.0",
+    "vue-cookie-law": "1.12.0",
     "vue2-google-maps": "^0.10.6"
   },
   "devDependencies": {
-    "@nuxtjs/eslint-config": "^2.0.0",
-    "@nuxtjs/eslint-module": "^1.1.0",
-    "babel-eslint": "^10.0.1",
-    "eslint": "^6.8.0",
-    "eslint-config-prettier": "^6.7.0",
-    "eslint-config-standard": ">=12.0.0",
-    "eslint-plugin-import": ">=2.16.0",
-    "eslint-plugin-jest": ">=22.3.0",
-    "eslint-plugin-node": ">=8.0.1",
-    "eslint-plugin-nuxt": ">=0.4.2",
-    "eslint-plugin-prettier": "^3.0.1",
-    "eslint-plugin-promise": ">=4.0.1",
-    "eslint-plugin-standard": ">=4.0.0",
-    "eslint-plugin-vue": "^6.0.1",
-    "nodemon": "^2.0.2",
-    "prettier": "^1.19.1"
+    "@nuxtjs/eslint-config": "^3.1.0",
+    "@nuxtjs/eslint-module": "^2.0.0",
+    "babel-eslint": "^10.1.0",
+    "eslint": "^7.8.1",
+    "eslint-config-prettier": "^6.11.0",
+    "eslint-plugin-nuxt": "^1.0.0",
+    "eslint-plugin-prettier": "^3.1.4",
+    "prettier": "^2.1.1"
   }
 }
